+++ front/guardias-web/src/app/components/listado-profesores/listado-profesores.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { HeaderComponent } from "../header/header.component";
 import { TableModule } from 'primeng/table';
-import { ProfesorService } from '../services/profesorService';
+import { ProfesorService } from '../../services/profesor.service';
 import { Button } from 'primeng/button';
 import { IconFieldModule } from 'primeng/iconfield';
 import { InputIconModule } from 'primeng/inputicon';
@@ -15,18 +15,14 @@ import { InputIconModule } from 'primeng/inputicon';
 })
 export class ListadoProfesoresComponent implements OnInit {
 
-  profesores: any[] | undefined;
+  profesores!: any[];
 
   rangeDates: Date[] | undefined;
-
-  prueba = [
-    { id: 1, nombre: 'Pepe', apellidos: 'Pérez', nif: '1234321', email: 'dfgdfg', telefono: '123123123' },
-  ]
 
   constructor(private ProfesorService: ProfesorService) { }
 
   ngOnInit() {
-    this.ProfesorService.findAll().subscribe(profesores => {
+    this.ProfesorService.findAll().subscribe((profesores: any[]) => {
       this.profesores = profesores;
       console.log(profesores)
     });
