--- front/guardias-web/src/app/components/services/profesor.ts
@@ -1,8 +1,0 @@
-export interface Profesor {
-    id: number;
-    nombre: string;
-    apellidos: string;
-    nif: string;
-    email: string;
-    telefono: string;
-}
