+++ front/guardias-web/src/app/services/profesor.service.ts
@@ -1,0 +1,19 @@
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ProfesorService {
+  private apiUrl = 'http://localhost:8000/api/profesores';
+
+  constructor(private http: HttpClient) { }
+
+  findAll(): any {
+    const headers = new HttpHeaders({
+      'Authorization': 'Basic ' + btoa('admin:admin')
+    });
+
+    return this.http.get<any[]>(this.apiUrl, { headers });
+  }
+}
